requirements.txt
flask==2.0.3
flask-login==0.5.0
flask-sqlalchemy==2.5.1
flask-wtf==1.0.1
flask-migrate==3.1.0
flask-session==0.4.0
psycopg2-binary==2.9.5
python-dotenv==0.20.0
Jinja2==3.0.3
itsdangerous==2.0.1
SQLAlchemy==1.4.46
gunicorn==20.1.0
email-validator==1.3.0
Werkzeug==2.0.3
Markdown==3.4.3
requests==2.28.1
openai==1.67.0
